chore(cluster): de-dupe creation of nodeSecurityGroup & ingress rule
In general, the `nodeSecurityGroup` and `eksClusterIngressRule` resources are
better suited to live and be managed in the `NodeGroup` class instead of
also in the `Cluster` class, and minimizing the areas to configure these
resources is better in the long-run.

The `nodeSecurityGroup` and its related `eksClusterIngressRule` are being
removed from the `Cluster` class, as these steps are
[duplicated and handled](https://github.com/pulumi/pulumi-eks/blob/master/nodejs/eks/nodegroup.ts#L245-L260) in
the `NodeGroup` class already if a user does not supply them.

Further example:

If a user tries to use the `Cluster` method `eks.Cluster.createNodeGroup()`,
and the user also supplies tags for the `nodeSecurityGroup`:
  - The user could expect to provide the tags in either the `Cluster` or
  `NodeGroup` class, and this could cause confusion.
  - If the user sets `skipDefaultNodeGroup: true` and supplies the
  `nodeSecurityGroup` tags through the `NodeGroup` and not the `Cluster`,
  these tags will not be applied. This is because the cluster's
  `nodeSecurityGroup` is being created in the `Cluster`, and
  passed as an arg to the `NodeGroup` using the API call in
  `eks.Cluster.createNodeGroup()`, and the `NodeGroup` tags can't be
  added on a readonly `tags` prop of the `nodeSecurityGroup`.

By removing both resources from the `Cluster` class, we'll simplify
working with the `nodeSecurityGroup` and its related rules, and make the
`NodeGroup` path and its specific resources more straight-forward.