chore(cluster): de-dupe creation of nodeSecurityGroup & ingress rule
The `nodeSecurityGroup` and `eksClusterIngressRule` for the
`nodeSecurityGroup` traffic into the `clusterSecurityGroup` are being
removed from the `Cluster` class, as these steps are
[duplicated](https://github.com/pulumi/pulumi-eks/blob/master/nodejs/eks/nodegroup.ts#L245-L260) in
the `NodeGroup` class already if a user does not supply them.

In general, the `nodeSecurityGroup` and `eksClusterIngressRule` resources are
better suited to live and be managed in the `NodeGroup` class instead of
also in the `Cluster` class, and minimizing the areas to configure these
resources is better in the long-run

Example: if a user tries to use the `Cluster` method
`eks.Cluster.createNodeGroup()`, and the user also supplies tags for the `nodeSecurityGroup`:

- The user would expect to provide the tags in either the `Cluster` or
`NodeGroup` class, and this could cause confusion.
- If the user sets `skipDefaultNodeGroup: true` and supplies the tags
through the `NodeGroup` these tags will not be applied, as the `Cluster`
`nodeSecurityGroup` was being created and passed as an arg to the NodeGroup`
via the API call in `eks.Cluster.createNodeGroup()`.

By removing both resources from the `Cluster` class, we'll simplify
working with the `nodeSecurityGroup` and related rules, and make working with
`NodeGroup` and its resources more straight-forward.