upgrades(eks): drain the t3.2xlarge node group
After draining, we must wait for the nginx-ing-cntlr instances to be drained
from its NLB service, and target group. This takes 5 minutes per the NLB
deregistration delay. We must wait for all drained k8s instances to be
deregistered from the NLB, before we can delete the nodes from the k8s APIServer.

https://docs.aws.amazon.com/elasticloadbalancing/latest/network/load-balancer-target-groups.html#deregistration-delay